Create the subsidiary table.
Creating the appropriate entry in pg_depend requires superuser
privilege, so that means a SECURITY DEFINER function. The version here
works correctly (and the ord does get dropped along with the relevant
column) but doesn't do any security checks.

I *think* what I need to to is check that ord and table have the same
owner, and that the session_user is a member of that role, but it's not
clear to me if that does the right thing if the user is already running
setid. I'd like to see what the sequence code does here, or maybe see if
there are existing extensions that try the same tricks.

Maybe I need a wrapper that records the current_user before going setid?
As long as I then verify that the session_user is a member of the passed
in role this shouldn't allow anyone to do anything they couldn't do with
SET ROLE, but keeps the advantage of inactive role memberships.