OK, so I'm stupid.
The bitstring implementation has one fatal flaw: before(o) always
returns the same value, so it's impossible to insert a new value between
two others. It is impossible in general for before() to be immutable, so
stop trying to store the ordering information in the value itself and
keep a proper tree in a separate table instead.

GCing the tree table will be tricky. Geting the entire table to
self-destruct when the column referencing it is dropped means making
each tree its own table rather than keeping them all in one; this in
turn means all the queries have to go through EXECUTE, which is less
than ideal. Clearing out old entries in the tree when the rows
referencing them are deleted will need a trigger on the referencing
table; this will make it impossible to keep a copy of an ordered value
in a PL/pgSQL variable, and any 'ordered' columns will need to be
UNIQUE.

I believe that both those problems could be fixed by using a custom
index method, since Pg has explicit support for correctly GCing index
rows, but I'd like to get an SQL version working first.