Ordering-creation functions.
Rename create_tree -> create_ordering. Revive set_ordering_for, but
rename it to _set_ordering_for and change it to just do the pg_depend
manipulations. The real set_ordering_for will do the rest: I think just
setting up constraints, but it may prove necessary to have a single-row
parameters table in addition to the tree, in which case it can record
the owner in there.

Still no permission-checking, so this function is currently rather
dangerous.