Check permissions when creating pg_depend entries.
This isn't perfect: there doesn't appear to be any way for a SECURITY
DEFINER function to find the current_user just before it was called, so
users will be allowed to do some things they aren't currently wearing
the right hat for. I also don't make exceptions for superusers, on the
grounds that if they're messing around they can make their own pg_depend
entries, thank-you-very-much :).

This should also really make some attempt to check that the ordering
actually is an ordering, and probably that the rel is a table and the
attribute has the right type. It may also be worth putting a 'n' dep on
the ordering from the ordered1 schema, since dropping the schema will
(probably) leave the ordering table useless.