FK triggers.
These triggers keep the ordering and its associated table column in
sync (they are almost but not quite entirely unlike a cascading FK).

ordered_insert should be BEFORE (probably) INSERT OR UPDATE on the table
column, and is essentially a cross-table CHECK.

ordered_delete and ordering_insert should be AFTER UPDATE OR DELETE on
the table and AFTER INSERT on the ordering, in both cases as an
INITIALLY DEFERRED constraint trigger. They both just call _ordered_gc,
which checks that the relevant ordered value did in fact end up back in
the table at the end of the txn, and deletes it if it didn't.