The planner plans the recursive with rather badly.
It seems not to be designed to cope with more than about 10 iterations
of the WITH, and to expect that each iteration would return a lot of
rows. Since exactly the opposite is true, the planner's use of seqscans
rather than index lookups means performance is appalling.

Hack around the problem by locally disabling seqscans. The right answer
is to use a btree instead of a plain binary tree, so the WITH doesn't
have to recurse so far. (A database table is, after all, secondary
storage.)