A btree implementation.
(Let's stop messing around with \ef and get this out into git.)

Delete works right for a single process, but ignores concurrency. Insert
needs to be told which node it's inserting into, and if that isn't a
leaf it doesn't behave.

I believe that, since I never use this tree for search, I want to move
all the keys down to the leaf level. This give me

    (id serial, kids integer[], leaf boolean)

where kids holds node pointers for non-leaf nodes and keys for leaf
nodes. Getting the locking right will be tricky, though.