Use a metapage to hold a pointer to the root.
While it is possible to find the root from the kids lists, it isn't
possible to use the index so it always requires a complete table scan.
Declare id = 0 to be a metapage holding a pointer to the root and the
fill size for this btree. Also declare that there is always a root page,
even if it's an empty leaf, since this simplifies things.